test(hacks): split return-split tests into own file for LOC hygiene
multimodal_test.go hit 874 lines, over the 800-line check-file-size.sh
gate. Move the self-contained return-split test block (tests + their
returnSplit* helpers) into multimodal_return_split_test.go, mirroring the
source file multimodal_return_split.go. Pure move: no test logic changed,
suite green under -race. Original drops to 270 lines, new file 618.
